move utils to tests package
