+++ common/djangoapps/third_party_auth/samlproviderdata/tests/test_samlproviderdata.py
@@ -15,7 +15,7 @@ from enterprise.constants import ENTERPRISE_ADMIN_ROLE
 
 from third_party_auth.tests import testutil
 from third_party_auth.models import SAMLProviderData, SAMLProviderConfig
-from third_party_auth.samlutils.utils import set_jwt_cookie
+from third_party_auth.tests.samlutils import set_jwt_cookie
 
 SINGLE_PROVIDER_CONFIG = {
     'entity_id': 'http://entity-id-1',
